1021-kadai15

+++ app/Http/Controllers/TasksController.php
@@ -16,13 +16,28 @@ class TasksController extends Controller
     //getでtasks/にアクセスされた場合の「一覧表示機能」
     public function index()
     {
-        //メッセージ一覧を取得
-        $tasks = Task::all();
+        
+        $data = [];
+        
+        //認証を実行
+        if(\Auth::check()){
+            
+            //認証済みユーザを取得
+            $user = \Auth::user();
+            
+            //ユーザのタスク一覧を取得 (１ページにつき10件)
+            $tasks = $user->tasks()->paginate(10);
+            
+            $data = [
+                "user" => $user,
+                "tasks" => $tasks,
+            ];
+            
+        }
+        
         
         //タスク一覧ビューでそれを表示
-        return view("tasks.index",[
-            "tasks"=> $tasks,
-        ]);
+        return view("welcome",$data);
     }
 
     /**
@@ -65,7 +80,15 @@ class TasksController extends Controller
         $task->status = $request->status;
         $task->user_id = \Auth::id();
         $task->content = $request->content;
-        $task->save();
+        //$task->save();
+        
+        
+        //認証済みユーザのタスクとして作成
+        $request->user()->tasks()->create([
+            "status" => $request->status,
+            "content" => $request->content,
+        ]);
+        
         
         //トップメニューへリダイレクトされる
         return redirect("/");
@@ -84,20 +107,10 @@ class TasksController extends Controller
          //idの値でメッセージを検索して取得
             $task = Task::findOrFail($id);
         
-
-        //認証済みユーザ（閲覧者）がそのタスクの登録者である場合は、投稿を閲覧
-        if(\Auth::id() === $task->user_id){
-            
-           
-        
             //メッセージ詳細ビューでそれを表示
             return view("tasks.show",[
                 "task" => $task,
             ]);
-        }
-        else{
-           return redirect("/"); 
-        }
     }
 
     /**
@@ -112,10 +125,17 @@ class TasksController extends Controller
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
         
-        //メッセージ編集ビューでそれを表示
-        return view("tasks.edit",[
-            "task"=> $task,
-        ]);
+        //認証済みユーザ（閲覧者）がそのタスクの登録者である場合は、投稿を閲覧
+        if(\Auth::id() === $task->user_id){
+        
+            //メッセージ編集ビューでそれを表示
+            return view("tasks.edit",[
+                "task"=> $task,
+            ]);
+        }
+        else{
+           return redirect("/"); 
+        }
     }
 
     /**
@@ -141,13 +161,17 @@ class TasksController extends Controller
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         
-        //タスクを更新  
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        //認証済みユーザ（閲覧者）がそのタスクの登録者である場合は、投稿を閲覧
+        if(\Auth::id() === $task->user_id){
+        
+            //タスクを更新  
+            $task->status = $request->status;
+            $task->content = $request->content;
+            $task->save();
         
         //トップページへリダイレクトさせる
-        return redirect("/");
+             return redirect("/");
+        }
     }
 
     /**
@@ -162,10 +186,15 @@ class TasksController extends Controller
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         
-        //メッセージを削除  
-        $task->delete();
         
-        //トップページへリダイレクトさせる
-        return redirect("/");
+        //認証済みユーザ（閲覧者）がそのタスクの登録者である場合は、投稿を閲覧
+        if(\Auth::id() === $task->user_id){
+        
+            //メッセージを削除  
+            $task->delete();
+            
+            //トップページへリダイレクトさせる
+            return redirect("/");
+        }
     }
 }
